3.Inheritance
Design and develop inheritance for a given case study, identify objects and relationships and implement
inheritance wherever applicable. Employee class with Emp_name, Emp_id, Address, Mail_id, and Mobile_no
as members. Inherit the classes, Programmer, Assistant Professor, Associate Professor and Professor from
employee class. Add Basic Pay (BP) as the member of all the inherited classes with 97% of BP as DA, 10 % of
BP as HRA, 12% of BP as PF, 0.1% of BP for staff club fund. Generate pay slips for the employees with their
gross and net salary.