Create 6.Exception handling
Implement a program to handle Arithmetic exception, Array Index Out Of Bounds.
The user enters two numbers Num1 and Num2. The division of Num1 and Num2 is displayed. If Num1 and
Num2 were not integers, the program would throw a Number Format Exception. If Num2 were zero, the
program would throw an Arithmetic Exception. Display the exception.